Fix ImageMagick command for v7 option handling
Options specified before the input image can be interpreted as options
to generate an image.

Example:

```
convert -size 300x200 xc:lightblue image.png
```

The `-size` and `xc:lightblue` options create a new image, and
`image.png` is considered the output file. 

By moving the options **after** the input image, we ensure that the
input file is interpreted correctly and only the output is modified
according to the specified options.

This will also add compatibility to ImageMagick v7